backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.0.0
alembic>=1.13.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.9
langchain-core>=0.2.0
tavily-python>=0.3.0
firecrawl-py>=0.0.16
python-dotenv>=1.0.0
pydantic>=2.0.0
sse-starlette>=1.8.0
email-validator>=2.0.0
langsmith>=0.1.0
